--- src/main/java/uk/gov/hmcts/reform/civil/notification/handlers/defendantResponse/unspec/offline/otherresponses/DefRespCaseOfflineAllLegalRepsEmailGenerator.java
@@ -1,19 +1,0 @@
-package uk.gov.hmcts.reform.civil.notification.handlers.defendantResponse.unspec.offline.otherresponses;
-
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.civil.notification.handlers.AllLegalRepsEmailGenerator;
-
-@Component
-public class DefRespCaseOfflineAllLegalRepsEmailGenerator extends AllLegalRepsEmailGenerator {
-
-    public DefRespCaseOfflineAllLegalRepsEmailGenerator(
-        DefRespCaseOfflineAppSolOneEmailDTOGenerator appSolOneEmailDTOGenerator,
-        DefRespCaseOfflineRespSolOneEmailDTOGenerator respSolOneEmailDTOGenerator,
-        DefRespCaseOfflineRespSolTwoEmailDTOGenerator respSolTwoEmailDTOGenerator
-    ) {
-        super(appSolOneEmailDTOGenerator,
-            respSolOneEmailDTOGenerator,
-            respSolTwoEmailDTOGenerator);
-    }
-
-}
