--- src/main/java/uk/gov/hmcts/reform/civil/notification/handlers/defendantResponse/unspec/offline/otherresponses/DefRespCaseOfflineUnspecNotifier.java
@@ -1,25 +1,0 @@
-package uk.gov.hmcts.reform.civil.notification.handlers.defendantResponse.unspec.offline.otherresponses;
-
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.civil.notification.handlers.Notifier;
-import uk.gov.hmcts.reform.civil.notification.handlers.adddefendantlitigantfriend.AddDefLitFriendAllLegalRepsEmailGenerator;
-import uk.gov.hmcts.reform.civil.notify.NotificationService;
-import uk.gov.hmcts.reform.civil.service.CaseTaskTrackingService;
-
-import static uk.gov.hmcts.reform.civil.notification.handlers.CamundaProcessIdentifier.DefendantResponseUnspecCaseHandedOfflineNotifyParties;
-
-@Component
-public class DefRespCaseOfflineUnspecNotifier
-    extends Notifier {
-
-    public DefRespCaseOfflineUnspecNotifier(NotificationService notificationService,
-                                            CaseTaskTrackingService caseTaskTrackingService,
-                                            AddDefLitFriendAllLegalRepsEmailGenerator allLegalRepsEmailGenerator) {
-        super(notificationService, caseTaskTrackingService, allLegalRepsEmailGenerator);
-    }
-
-    @Override
-    public String getTaskId() {
-        return DefendantResponseUnspecCaseHandedOfflineNotifyParties.toString();
-    }
-}
