--- src/main/java/uk/gov/hmcts/reform/civil/notification/handlers/defendantResponse/unspec/online/fulldefence/DefendantResponseAllLegalRepsEmailGenerator.java
@@ -1,18 +1,0 @@
-package uk.gov.hmcts.reform.civil.notification.handlers.defendantResponse.unspec.online.fulldefence;
-
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.civil.notification.handlers.AllLegalRepsEmailGenerator;
-
-@Component
-public class DefendantResponseAllLegalRepsEmailGenerator extends AllLegalRepsEmailGenerator {
-
-    public DefendantResponseAllLegalRepsEmailGenerator(
-        DefendantResponseAppSolOneEmailDTOGenerator appSolOneEmailDTOGenerator,
-        DefendantResponseRespSolOneEmailDTOGenerator respSolOneEmailDTOGenerator,
-        DefendantResponseRespSolTwoEmailDTOGenerator respSolTwoEmailDTOGenerator
-    ) {
-        super(appSolOneEmailDTOGenerator,
-            respSolOneEmailDTOGenerator,
-            respSolTwoEmailDTOGenerator);
-    }
-}
