+++ src/main/java/uk/gov/hmcts/reform/civil/notification/handlers/defendantresponse/unspec/offline/otherresponses/DefRespCaseOfflineHelper.java
@@ -1,0 +1,55 @@
+package uk.gov.hmcts.reform.civil.notification.handlers.defendantresponse.unspec.offline.otherresponses;
+
+import uk.gov.hmcts.reform.civil.model.CaseData;
+
+import java.util.Map;
+
+import static uk.gov.hmcts.reform.civil.enums.CaseCategory.SPEC_CLAIM;
+import static uk.gov.hmcts.reform.civil.enums.MultiPartyScenario.ONE_V_ONE;
+import static uk.gov.hmcts.reform.civil.enums.MultiPartyScenario.TWO_V_ONE;
+import static uk.gov.hmcts.reform.civil.enums.MultiPartyScenario.getMultiPartyScenario;
+import static uk.gov.hmcts.reform.civil.handler.callback.camunda.notification.NotificationData.REASON;
+import static uk.gov.hmcts.reform.civil.handler.callback.camunda.notification.NotificationData.RESPONDENT_ONE_NAME;
+import static uk.gov.hmcts.reform.civil.handler.callback.camunda.notification.NotificationData.RESPONDENT_ONE_RESPONSE;
+import static uk.gov.hmcts.reform.civil.handler.callback.camunda.notification.NotificationData.RESPONDENT_TWO_NAME;
+import static uk.gov.hmcts.reform.civil.handler.callback.camunda.notification.NotificationData.RESPONDENT_TWO_RESPONSE;
+import static uk.gov.hmcts.reform.civil.utils.PartyUtils.getPartyNameBasedOnType;
+
+public class DefRespCaseOfflineHelper {
+
+    private DefRespCaseOfflineHelper() {
+        //NO-OP
+    }
+
+    public static Map<String, String> caseOfflineNotificationProperties(CaseData caseData) {
+        if (getMultiPartyScenario(caseData).equals(ONE_V_ONE)) {
+            return Map.of(
+                REASON, caseData.getRespondent1ClaimResponseType().getDisplayedValue()
+            );
+        } else if (getMultiPartyScenario(caseData).equals(TWO_V_ONE)) {
+            String responseTypeToApplicant2 = SPEC_CLAIM.equals(caseData.getCaseAccessCategory())
+                ? caseData.getClaimant1ClaimResponseTypeForSpec().getDisplayedValue()
+                : caseData.getRespondent1ClaimResponseTypeToApplicant2().toString();
+            return Map.of(
+                REASON, SPEC_CLAIM.equals(caseData.getCaseAccessCategory())
+                    ? caseData.getClaimant1ClaimResponseTypeForSpec().getDisplayedValue()
+                    : caseData.getRespondent1ClaimResponseType().getDisplayedValue()
+                    .concat(" against " + caseData.getApplicant1().getPartyName())
+                    .concat(" and " + responseTypeToApplicant2)
+                    .concat(" against " + caseData.getApplicant2().getPartyName())
+            );
+        } else {
+            //1v2 template is used and expects different data
+            return Map.of(
+                RESPONDENT_ONE_NAME, getPartyNameBasedOnType(caseData.getRespondent1()),
+                RESPONDENT_TWO_NAME, getPartyNameBasedOnType(caseData.getRespondent2()),
+                RESPONDENT_ONE_RESPONSE, SPEC_CLAIM.equals(caseData.getCaseAccessCategory())
+                    ? caseData.getRespondent1ClaimResponseTypeForSpec().getDisplayedValue()
+                    : caseData.getRespondent1ClaimResponseType().getDisplayedValue(),
+                RESPONDENT_TWO_RESPONSE, SPEC_CLAIM.equals(caseData.getCaseAccessCategory())
+                    ? caseData.getRespondent2ClaimResponseTypeForSpec().getDisplayedValue()
+                    : caseData.getRespondent2ClaimResponseType().getDisplayedValue()
+            );
+        }
+    }
+}
